Add Terraform config for VPC and autoscaling group
The VPC has a NAT that allows the EC2 instances in the private subnets
to be able to connect to the ECS cluster. I think the NAT gateway will
cost some small amount of money per hour.

The autoscaling group spawns the EC2 instance(s) that will connect to
the ECS cluster and will eventually run the actual services/tasks.